requirements.txt
alembic==1.9.1
bidict==0.22.1
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-MySQL==1.5.2
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
protobuf==3.20.3
PyMySQL==1.0.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
simple-websocket==0.9.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
wsproto==1.2.0
